+++ src/components/TextInput/index.ts
@@ -1,2 +1,2 @@
 export { default as TextInput } from './TextInput.vue'
-export type { TextInputProps } from './types'
+export type { TextInputExposed, TextInputProps } from './types'